[BOJ-2161] ì카드1 코ëㅍ일 수정
